+++ back/src/common/redis/index.ts
@@ -1,0 +1,1 @@
+export * from "./redis";
